+++ pkg/gnubuilder/flags.go
@@ -1,4 +1,4 @@
-package builder
+package gnubuilder
 
 import "regexp"
